Tentative display class. Low level interface
Havn't decided on a graphics API to use. So most of this is tentative. Again this is a low level interface, which can then be passed off to the graphics API for it to be draw. 

Most of the functions here will do "chunkalating" of the display space into a grid where tiles could be draw.