Added food quote parser. Uses ugly popup display.
Added in a parser for the food quotes.

Moved the display from the log to a popup window.  It is very hideous at
the moment though.  We'll have to find a good panel Layout.  I tried
Grid, Flow, GridBag, DefaultMenu and a few others, but they all looked
bad.  If anybody knows how to make a pretty popup, let me know.